% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsfit_diagnostics.R
\name{check_brmsfit_diagnostics}
\alias{check_brmsfit_diagnostics}
\title{check brmsfit diagnostics}
\usage{
check_brmsfit_diagnostics(brmsfit)
}
\arguments{
\item{brmsfit}{a brmsfit object}
}
\value{
logical vector. First element is whether divergences passed; second
is whether rhat passed
}
\description{
check brmsfit diagnostics
}
