% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daymet_extract.R
\name{daymet_set_extract}
\alias{daymet_set_extract}
\title{Extract a set of Daymet parameters for a cell and year}
\usage{
daymet_set_extract(year, cell, grid, params_daymet)
}
\arguments{
\item{year}{the year of interest}

\item{cell}{the large grid index}

\item{grid}{a grid specification obtained via \link[dggridR]{dgconstruct}}

\item{params_daymet}{a data.frame specifying daymet parameters and averaging periods}
}
\value{
a data.frame with cell-averaged daymet parameters
}
\description{
Extracts a set of Daymet data according to a specification of variables, minimum and maximum dates
using \link{daymet_extract}
}
\details{
Parameter \code{params_daymet} should be a \code{data.frame} with the following columns:
\describe{
\item{\code{label}}{a user-defined label to define the entry}
\item{\code{variable}}{a valid Daymet variable}
\item{\code{date_min}}{a minimum date formatted as mm-dd (without year)}
\item{\code{date_max}}{a maximum date formatted as mm-dd (without year)}
\item{\code{period}}{the corresponding demographic period (to associate a regression using \link{weather_regressions})}
}
}
