% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_and_bootstrap.R
\name{get_abun}
\alias{get_abun}
\title{Get abundance data}
\usage{
get_abun(sp_data, n_rep, roi, .cores = 4)
}
\arguments{
\item{sp_data}{Data for a particular species and year from get_grid_data()}

\item{n_rep}{number of bootstrap replicates}

\item{roi}{a raster object defining the Region Of Interest (ROI)}

\item{.cores}{number of parallel cores to use}
}
\value{
a dataframe
}
\description{
Get abundance data
}
