% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_timeseries.R
\name{get_abun_summary}
\alias{get_abun_summary}
\title{Abundance summary across years}
\usage{
get_abun_summary(abun_data, n_small_min)
}
\arguments{
\item{abun_data}{abundance data}

\item{n_small_min}{minimum number of small cells to compute abundance index for large cell}
}
\value{
a list of data.frames for each year calculated by \code{\link[=abun_data_bycell]{abun_data_bycell()}}.
}
\description{
Abundance summary across years
}
\details{
Each list element for this function contains a data.frame for a year, provided by \code{abun_data}
}
