% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_and_bootstrap.R
\name{get_abun_tgrid_slice}
\alias{get_abun_tgrid_slice}
\title{Get abundance for one timeslice}
\usage{
get_abun_tgrid_slice(sp_data, t, cells, big_cells, n_rep)
}
\arguments{
\item{sp_data}{Data for a particular species and year from get_grid_data()}

\item{t}{index for the timeslice}

\item{cells}{output of get_pixels() subset to ROI}

\item{big_cells}{list of the big-cell identities}

\item{n_rep}{number of bootstrap replicates}
}
\description{
Get abundance for one timeslice
}
