% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_timeseries.R
\name{get_cell_timeseries}
\alias{get_cell_timeseries}
\title{Merge spring and fall indices into a timeseries}
\usage{
get_cell_timeseries(cells_all, spring_abun_summary, fall_abun_summary)
}
\arguments{
\item{cells_all}{numeric vector of cell ids}

\item{spring_abun_summary}{spring abundance summary}

\item{fall_abun_summary}{fall abundance summary}
}
\value{
a list of data.frames, with list elements corresponding to cells
}
\description{
Merge spring and fall indices into a timeseries
}
\details{
This functions pastes the spring and full abundance summaries together
in one data.frame for each cell, with the spring on the uneven column indices and fall on the
even column indices, with in total 2x (number of years) columns
}
