% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_data.R
\name{get_grid_data}
\alias{get_grid_data}
\title{Get eBird summaries on cells of a small grid, tagged with membership in a large grid.}
\usage{
get_grid_data(
  data,
  .year,
  tgrid_min,
  tgrid_max,
  time_window = "gridded",
  min_lat,
  max_lat,
  min_lon,
  max_lon = Inf,
  large_grid = 6,
  small_grid = 11,
  time_grid = 7,
  roi,
  .cores = 4
)
}
\arguments{
\item{data}{zero-filled data (output from import_from_erd())}

\item{.year}{the year}

\item{tgrid_min}{the minimum time grid index (week number if \code{time_grid=7}).}

\item{tgrid_max}{the maximum time grid index (week number if \code{time_grid=7}).}

\item{time_window}{"gridded" or "full"}

\item{min_lat}{the minimum latitude (decimal degrees) where the prediction is desired}

\item{max_lat}{the maximum latitude}

\item{min_lon}{the minimum longitude}

\item{max_lon}{the maximum longitude}

\item{large_grid}{the dggridR resolution of the large grid}

\item{small_grid}{the dggridR resolution of the small grid}

\item{time_grid}{the temporal grid size in days (default 7, one week).}

\item{roi}{region of interest}

\item{.cores}{number of parallel cores for \code{get_cell_data()} calls}
}
\value{
a list with grid-sampled data
}
\description{
Get eBird summaries on cells of a small grid, tagged with membership in a large grid.
}
\details{
The returned list object either contains a single list of lists (when time_window equals full)
or a multiple lists of lists (when time_window equals weekly), one for each week. The deepest
list elements contains count information for the small grid, with list elements labled as
\code{cell_X} with \code{X} the small cell index. These small cell data elements contain the following
list elements returned by \code{\link[=get_cell_data]{get_cell_data()}}:
\describe{
\item{\code{n}}{number of observations (number of zero-filled checklist available in the small cell)}
\item{\code{n_z}}{number of zero counts}
\item{\code{n_x}}{number of presence only observations (X's)}
\item{\code{n_value}}{number of observations with a count value (zero or a count, i.e. non-X observations)}
\item{\code{mean_positive}}{average count of observations with a count > 0}
\item{\code{stixel_mean_small}}{average count for the small cell, as calculated with \code{\link[=stixel_mean_small]{stixel_mean_small()}}}
}
The average count calculated with \code{\link[=stixel_mean_small]{stixel_mean_small()}} currently assumes for presence only observations
(X's) a count equal to \code{mean_positive}.
}
