% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_grid_abun.R
\name{get_higher_moments}
\alias{get_higher_moments}
\title{Get higher moments of log ratios}
\usage{
get_higher_moments(cell_ratios, cells_all, cells)
}
\arguments{
\item{cell_ratios}{cell ratios object}

\item{cells_all}{names of all cells underlying cell_ratios}

\item{cells}{all cells over which output is desired}
}
\value{
list giving all skews and all excess kurtoses
}
\description{
Get higher moments of log ratios
}
