% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_data.R
\name{get_pixels}
\alias{get_pixels}
\title{get a hexagonal grid over the contiguous US}
\usage{
get_pixels(dg_large, dg_small, roi)
}
\arguments{
\item{dg_large}{large dggridR grid}

\item{dg_small}{small dggridR grid}

\item{roi}{a raster object defining the Region Of Interest (ROI)}
}
\description{
get a hexagonal grid over the contiguous US
}
