% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratios.R
\name{get_ratios}
\alias{get_ratios}
\title{Get log-ratios}
\usage{
get_ratios(
  data,
  cells_all,
  period = c("spring", "fall"),
  n_small_min = 10,
  quiet = TRUE
)
}
\arguments{
\item{data}{abundance index data}

\item{cells_all}{cells to compute}

\item{period}{period to compute}

\item{n_small_min}{minimum number of small cells required to calculate an abundance ratio for a large cell}

\item{quiet}{if TRUE, suppress informational messages (warnings/errors still returned)}
}
\description{
Get log-ratios
}
