% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_data.R
\name{get_tgrid}
\alias{get_tgrid}
\title{get a 7-day grid over ordinal days}
\usage{
get_tgrid(days = 7)
}
\arguments{
\item{days}{time grid resolution in days}
}
\value{
a two-column data-frame that tags each day-of-year to a tgrid cell
}
\description{
get a 7-day grid over ordinal days
}
