% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_erd.R
\name{import_checklists}
\alias{import_checklists}
\title{Import checklists data from the erd}
\usage{
import_checklists(
  erd_path = "/Users/jacobsocolar/Dropbox/Work/macrodemography/erd/erd.db"
)
}
\arguments{
\item{erd_path}{The path to the eBird reference dataset .db file}
}
\value{
a data.frame of the checklists
}
\description{
Import checklists data from the erd
}
