% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_erd.R
\name{import_from_erd}
\alias{import_from_erd}
\title{Import and zero-fill data from the ebird reference dataset}
\usage{
import_from_erd(
  sp_code,
  erd_path = "/Users/jacobsocolar/Dropbox/Work/macrodemography/erd/erd.db",
  checklists = NULL,
  obs = NULL
)
}
\arguments{
\item{sp_code}{A six-letter species code from eBird (e.g. "tenwar")}

\item{erd_path}{The path to the eBird reference dataset .db file}

\item{checklists}{An optional path to a previously extracted checklists file}

\item{obs}{An optional path to a previously extracted observation file}
}
\value{
A \code{data.table} object with zero-filled abundance and presence-only reported colums
as well as additional columns for latitude, longitude, year, month, day_of_year, hours_of_day,
protocol_id, is_stationary, is_traveling, effort_hrs, effort_distance_km, cci
}
\description{
Import and zero-fill data from the ebird reference dataset
}
