% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrodemography.R
\docType{package}
\name{macrodemography-package}
\alias{macrodemography}
\alias{macrodemography-package}
\title{macrodemography: Functions for summarizing abundance from ERD}
\description{
Summarize ERD in nice ways.
}
\details{
Estimation of Macrodemographic indices from eBird DAta
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
