% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ratios_tidy.R
\name{make_ratios_tidy}
\alias{make_ratios_tidy}
\title{reformat the output of get_ratios() into tidy format}
\usage{
make_ratios_tidy(cell_ratios)
}
\arguments{
\item{cell_ratios}{data output from \code{\link[=get_ratios]{get_ratios()}}}
}
\value{
a list with two tidy data frames
}
\description{
reformat the output of get_ratios() into tidy format
}
