% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{plot_ns_summary}
\alias{plot_ns_summary}
\title{plotting function for two summaries returned by summarize_avg_and_rep,
one for a northern region and one for a southern region.}
\usage{
plot_ns_summary(results_summary_n, results_summary_s)
}
\arguments{
\item{results_summary_n}{summary for the northern region}

\item{results_summary_s}{summary for the southern region}
}
\description{
plotting function for two summaries returned by summarize_avg_and_rep,
one for a northern region and one for a southern region.
}
