% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{plot_one_summary}
\alias{plot_one_summary}
\title{plotting function for summary returned by summarize_avg_and_rep}
\usage{
plot_one_summary(
  results_summary,
  speed_data_spring = NULL,
  speed_data_fall = NULL
)
}
\arguments{
\item{results_summary}{summary returned by summarize_avg_and_rep}

\item{speed_data_spring}{spring speed data from median_passage}

\item{speed_data_fall}{fall speed data from median_passage}
}
\description{
plotting function for summary returned by summarize_avg_and_rep
}
