% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratios.R
\name{plot_ratios}
\alias{plot_ratios}
\title{plot the demographic ratio data for a large cell}
\usage{
plot_ratios(cell_index, data, log = TRUE)
}
\arguments{
\item{cell_index}{the index of a large cell}

\item{data}{a tidy dataframe of ratio data, typically output from \link{make_ratios_tidy}}

\item{log}{whether to plot ratios on a logarithmic scale}
}
\description{
plot the demographic ratio data for a large cell
}
