% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_test.R
\name{variance_test}
\alias{variance_test}
\title{perform formal test of whether productivity or survival variance is larger}
\usage{
variance_test(cell_ratios, cells_all, cells, years, min_n = 5)
}
\arguments{
\item{cell_ratios}{cell ratios object}

\item{cells_all}{all cells in cell_ratios}

\item{cells}{the cells over which output is desired}

\item{years}{years underlying the cell ratios object}

\item{min_n}{minimum sample size for which to estimate a variance}
}
\value{
dataframe. prob_surv gives posterior probability that survival variance
is higher than productivity variance
coef is the average log-scale effect.
}
\description{
perform formal test of whether productivity or survival variance is larger
}
