% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_regressions.R
\name{weather_regressions}
\alias{weather_regressions}
\title{perform weather regressions. currently expects to see data (and perform regressions)
for winter temp, winter swe, and summer temp, but this would be easily updated by
setting NULL defaults for these inputs plus a bit of control flow.}
\usage{
weather_regressions(
  cell_ratios,
  cells_all,
  years,
  min_n,
  janfeb_temp,
  julaug_temp,
  decmar_swe
)
}
\arguments{
\item{cell_ratios}{cell ratios object}

\item{cells_all}{all cells in cell_ratios}

\item{years}{years underlying the cell ratios object}

\item{min_n}{minimum sample size for which to estimate a slope}

\item{janfeb_temp}{jan/feb temps}

\item{julaug_temp}{july/august temps}

\item{decmar_swe}{dec-mar swe}
}
\value{
2-element list.  prob_surv gives posterior probability that survival variance
is higher than productivity variance
coef is the average log-scale effect.
}
\description{
perform weather regressions. currently expects to see data (and perform regressions)
for winter temp, winter swe, and summer temp, but this would be easily updated by
setting NULL defaults for these inputs plus a bit of control flow.
}
