% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_test.R
\name{compare_ratio_variances}
\alias{compare_ratio_variances}
\title{perform formal test of whether productivity or survival variance is larger}
\usage{
compare_ratio_variances(
  cell_index,
  data,
  n_ratio_min = 5,
  warmup = 1000,
  iter = 2000,
  chains = 3
)
}
\arguments{
\item{cell_index}{cell index to compare ratios for}

\item{data}{a tidy data.frame with ratio data, typically output from \link{make_ratios_tidy}}

\item{n_ratio_min}{minimum number of seasonal ratios required to calculate a variance}

\item{warmup}{number of warmup samples}

\item{iter}{number of samples to characterize the posterior distribution}

\item{chains}{number of parallel mcmc chains.}
}
\value{
dataframe. p_survival_variance_higher gives posterior probability that survival variance
is higher than productivity variance. effect_size_log is the average log-scale effect.
}
\description{
perform formal test of whether productivity or survival variance is larger
}
