% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_regressions.R
\name{weather_regressions}
\alias{weather_regressions}
\title{Perform weather regressions.}
\usage{
weather_regressions(
  tidy_ratios,
  data_daymet,
  params_daymet,
  min_n,
  warmup = 1000,
  iter = 2000,
  chains = 3,
  quiet = FALSE
)
}
\arguments{
\item{tidy_ratios}{ratios object, typically output of \link{make_ratios_tidy}}

\item{data_daymet}{weather data, typically output of \link{daymet_set_extract}}

\item{params_daymet}{data.frame specifying Daymet data extraction, see Details section of \link{daymet_set_extract}}

\item{min_n}{minimum sample size for which to estimate a slope}

\item{warmup}{number of warmup samples}

\item{iter}{number of samples to characterize the posterior distribution}

\item{chains}{number of parallel mcmc chains.}

\item{quiet}{if TRUE, suppress informational messages (warnings/errors still returned)}
}
\value{
a data.frame with regression results
}
\description{
Perform weather regressions.
}
